+++ supportCodeLibrary/test_run_hook_definition.js
@@ -1,3 +1,3 @@
-import Definition from './definition'
+import Definition from './definition.js'
 
 export default class TestRunHookDefinition extends Definition {}